# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Blog Api",
        "version": "1"
    },
    "paths": {
        "/api/author/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Find by ID",
                "description": "Returns a single author",
                "operationId": "d4fb0358eb484f36f719ae61f5d589bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of author to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Author with id 3 retrieved"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/author/login": {
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Login",
                "description": "Performs login attempt and retrieves access token if success",
                "operationId": "c9b880bd175fd6021527e4ccd55798fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "winona78@example.net",
                                "password": "password"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login success"
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "1|gXppSgPS7E2UMGF8jsMGJuM7FM5MSYH2BAfmmSsT"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Find by ID",
                "description": "Returns a single post",
                "operationId": "d31af682add10e50941f21e84f36d436",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of post to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Author with id 3 retrieved"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/list": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List of posts",
                "description": "Get complete list of posts with pagination options",
                "operationId": "d9c55bd0f40e49e0e0878893d1035501",
                "parameters": [
                    {
                        "name": "results",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Amount of post to be retrieved (by default all), starting by order field (default id) and ignoring first n posts specified by offset parameter (default 0)",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Ignore first n posts (by default 0)",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_field",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Determines field to order results (by default id)",
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_ascending",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Determines if order is ascending (default) or descending",
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List retrieved"
                                        },
                                        "payload": {
                                            "properties": {
                                                "total_posts": {
                                                    "type": "integer",
                                                    "example": "50"
                                                },
                                                "retrieved_posts": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Post"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/post/create": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create post",
                "description": "Creates a new post and retrieves created post ID",
                "operationId": "b08a1fa600c97d746857352d3f9c97bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Post"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "new-post"
                                    },
                                    "extract": {
                                        "type": "string",
                                        "example": "Short text to describe the post"
                                    },
                                    "post": {
                                        "type": "string",
                                        "example": "<p>Text of the post which accepts HTML tags</p>"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post created"
                                        },
                                        "payload": {
                                            "type": "string",
                                            "example": "51"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authorization failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Author",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "name": {
                        "type": "string",
                        "example": "Great Author"
                    },
                    "email": {
                        "type": "string",
                        "example": "great@author.au"
                    },
                    "post_count": {
                        "type": "integer",
                        "example": "5"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "title": {
                        "type": "string",
                        "example": "Aperiam quis nisi laboriosam."
                    },
                    "slug": {
                        "type": "string",
                        "example": "voluptatem-quos-magni-nam-sunt-magni-tenetur"
                    },
                    "post": {
                        "type": "string",
                        "example": "<p>HTML content</p>"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "extract": {
                        "type": "string",
                        "example": "Omnis quas at est sequi maiores maiores quis debitis. Sed quis et nihil neque quibusdam ut. Harum labore aut dolores. Tempora eveniet natus voluptates error aliquid ut id."
                    },
                    "visits": {
                        "type": "integer",
                        "example": "3"
                    },
                    "created_at": {
                        "type": "date-time",
                        "example": "2023-08-06 18:49:58"
                    },
                    "updated_at": {
                        "type": "date-time",
                        "example": "2023-08-06 18:49:58"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token generated by login function",
                "name": "Sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}